Move notifications to its own tab
