+++ app/src/models/preferences.ts
@@ -1,8 +1,9 @@
 export enum PreferencesTab {
-  Accounts = 0,
-  Integrations = 1,
-  Git = 2,
-  Appearance = 3,
-  Prompts = 4,
-  Advanced = 5,
+  Accounts,
+  Integrations,
+  Git,
+  Appearance,
+  Notifications,
+  Prompts,
+  Advanced,
 }
