+++ app/src/ui/preferences/advanced.tsx
@@ -6,24 +6,13 @@ import { SamplesURL } from '../../lib/stats'
 import { UncommittedChangesStrategy } from '../../models/uncommitted-changes-strategy'
 import { RadioButton } from '../lib/radio-button'
 import { isWindowsOpenSSHAvailable } from '../../lib/ssh/ssh'
-import {
-  getNotificationSettingsUrl,
-  supportsNotifications,
-  supportsNotificationsPermissionRequest,
-} from 'desktop-notifications'
-import {
-  getNotificationsPermission,
-  requestNotificationsPermission,
-} from '../main-process-proxy'
 
 interface IAdvancedPreferencesProps {
   readonly useWindowsOpenSSH: boolean
   readonly optOutOfUsageTracking: boolean
-  readonly notificationsEnabled: boolean
   readonly uncommittedChangesStrategy: UncommittedChangesStrategy
   readonly repositoryIndicatorsEnabled: boolean
   readonly onUseWindowsOpenSSHChanged: (checked: boolean) => void
-  readonly onNotificationsEnabledChanged: (checked: boolean) => void
   readonly onOptOutofReportingChanged: (checked: boolean) => void
   readonly onUncommittedChangesStrategyChanged: (
     value: UncommittedChangesStrategy
@@ -35,9 +24,6 @@ interface IAdvancedPreferencesState {
   readonly optOutOfUsageTracking: boolean
   readonly uncommittedChangesStrategy: UncommittedChangesStrategy
   readonly canUseWindowsSSH: boolean
-  readonly suggestGrantNotificationPermission: boolean
-  readonly warnNotificationsDenied: boolean
-  readonly suggestConfigureNotifications: boolean
 }
 
 export class Advanced extends React.Component<
@@ -51,15 +37,11 @@ export class Advanced extends React.Component<
       optOutOfUsageTracking: this.props.optOutOfUsageTracking,
       uncommittedChangesStrategy: this.props.uncommittedChangesStrategy,
       canUseWindowsSSH: false,
-      suggestGrantNotificationPermission: false,
-      warnNotificationsDenied: false,
-      suggestConfigureNotifications: false,
     }
   }
 
   public componentDidMount() {
     this.checkSSHAvailability()
-    this.updateNotificationsState()
   }
 
   private async checkSSHAvailability() {
@@ -92,12 +74,6 @@ export class Advanced extends React.Component<
     event: React.FormEvent<HTMLInputElement>
   ) => {
     this.props.onUseWindowsOpenSSHChanged(event.currentTarget.checked)
-  }
-
-  private onNotificationsEnabledChanged = (
-    event: React.FormEvent<HTMLInputElement>
-  ) => {
-    this.props.onNotificationsEnabledChanged(event.currentTarget.checked)
   }
 
   private reportDesktopUsageLabel() {
@@ -162,7 +138,6 @@ export class Advanced extends React.Component<
           </p>
         </div>
         {this.renderSSHSettings()}
-        {this.renderNotificationsSettings()}
         <div className="advanced-section">
           <h2>Usage</h2>
           <Checkbox
@@ -195,106 +170,6 @@ export class Advanced extends React.Component<
           onChange={this.onUseWindowsOpenSSHChanged}
         />
       </div>
-    )
-  }
-
-  private renderNotificationsSettings() {
-    return (
-      <div className="advanced-section">
-        <h2>Notifications</h2>
-        <Checkbox
-          label="Enable notifications"
-          value={
-            this.props.notificationsEnabled
-              ? CheckboxValue.On
-              : CheckboxValue.Off
-          }
-          onChange={this.onNotificationsEnabledChanged}
-        />
-        <p className="git-settings-description">
-          Allows the display of notifications when high-signal events take place
-          in the current repository.{this.renderNotificationHint()}
-        </p>
-      </div>
-    )
-  }
-
-  private onGrantNotificationPermission = async () => {
-    await requestNotificationsPermission()
-    this.updateNotificationsState()
-  }
-
-  private async updateNotificationsState() {
-    const notificationsPermission = await getNotificationsPermission()
-    this.setState({
-      suggestGrantNotificationPermission:
-        supportsNotificationsPermissionRequest() &&
-        notificationsPermission === 'default',
-      warnNotificationsDenied: notificationsPermission === 'denied',
-      suggestConfigureNotifications: notificationsPermission === 'granted',
-    })
-  }
-
-  private renderNotificationHint() {
-    // No need to bother the user if their environment doesn't support our
-    // notifications or if they've been explicitly disabled.
-    if (!supportsNotifications() || !this.props.notificationsEnabled) {
-      return null
-    }
-
-    const {
-      suggestGrantNotificationPermission,
-      warnNotificationsDenied,
-      suggestConfigureNotifications,
-    } = this.state
-
-    if (suggestGrantNotificationPermission) {
-      return (
-        <>
-          {' '}
-          You need to{' '}
-          <LinkButton onClick={this.onGrantNotificationPermission}>
-            grant permission
-          </LinkButton>{' '}
-          to display these notifications from GitHub Desktop.
-        </>
-      )
-    }
-
-    const notificationSettingsURL = getNotificationSettingsUrl()
-
-    if (notificationSettingsURL === null) {
-      return null
-    }
-
-    if (warnNotificationsDenied) {
-      return (
-        <>
-          <br />
-          <br />
-          <span className="warning-icon">⚠️</span> GitHub Desktop has no
-          permission to display notifications. Please, enable them in the{' '}
-          <LinkButton uri={notificationSettingsURL}>
-            Notifications Settings
-          </LinkButton>
-          .
-        </>
-      )
-    }
-
-    const verb = suggestConfigureNotifications
-      ? 'properly configured'
-      : 'enabled'
-
-    return (
-      <>
-        {' '}
-        Make sure notifications are {verb} for GitHub Desktop in the{' '}
-        <LinkButton uri={notificationSettingsURL}>
-          Notifications Settings
-        </LinkButton>
-        .
-      </>
     )
   }
 }
